requirements.txt
bentoml==1.0.10
joblib==1.2.0
pip-chill==1.0.1
# torchvision==0.14.1
torchvision
tqdm==4.64.1
wincertstore==0.2
torchsummary==1.5.1
matplotlib==3.10.3
opencv-python==4.12.0.88
fastapi
uvicorn

cd Lungs-Disease-Diagnosis-master
env\Scripts\activate
(env) C:\Users\Dell\Desktop\lung\Lungs-Disease-Diagnosis-master>
uvicorn app:app --reload

